Rename Signature.ts to Signer.ts
