+++ src/crypto/Signer.test.ts
@@ -6,7 +6,7 @@
  * The above notice should be included in all copies or substantial portions of the software.
  */
 
-import * as Signature from "./Signature";
+import { signEd25519, verifyEd25519 } from "./Signer";
 import { getTextEncoder } from "membuffers";
 
 const someDataToSign = getTextEncoder().encode("this is what we want to sign");
@@ -14,9 +14,9 @@ const PublicKey1 = Buffer.from("92d469d7c004cc0b24a192d9457836bf38effa27536627ef
 const PrivateKey1 = Buffer.from("3b24b5f9e6b1371c3b5de2e402a96930eeafe52111bb4a1b003e5ecad3fab53892d469d7c004cc0b24a192d9457836bf38effa27536627ef60718b00b0f33152", "hex");
 
 test("SignEd25519", () => {
-  const sig = Signature.signEd25519(PrivateKey1, someDataToSign);
-  expect(Signature.verifyEd25519(PublicKey1, someDataToSign, sig)).toBe(true);
+  const sig = signEd25519(PrivateKey1, someDataToSign);
+  expect(verifyEd25519(PublicKey1, someDataToSign, sig)).toBe(true);
 
   sig[0] += 1; // corrupt the signature
-  expect(Signature.verifyEd25519(PublicKey1, someDataToSign, sig)).toBe(false);
+  expect(verifyEd25519(PublicKey1, someDataToSign, sig)).toBe(false);
 });
